+++ CHANGES.md
@@ -16,6 +16,7 @@ unreleased
     - Fix `construct` results ordering for sum types sand poly variants (#1603)
     - Fix object method completion not working (#1606, fixes #1575)
     - Improve context detection for package types (#1608, fixes #1607)
+    - Fix incorrect locations for string literals (#1574)
   + editor modes
     - emacs: call the user's configured completion UI in
       `merlin-construct` (#1598)